requirements.txt
altair==4.2.2
opencv-python-headless==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
streamlit==1.21.0
tensorflow==2.11.0
